refactor: revert some changes from 7a43b81
Moving `ImportMap` struct to `functions/pkg` and importing it inside
internals
