+++ pkg/function/deploy_test.go
@@ -14,11 +14,9 @@ import (
 	fs "testing/fstest"
 
 	"github.com/h2non/gock"
-	"github.com/spf13/afero"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
-	"github.com/supabase/cli/internal/utils"
 	"github.com/supabase/cli/pkg/api"
 	"github.com/supabase/cli/pkg/cast"
 	"github.com/supabase/cli/pkg/config"
@@ -51,7 +49,7 @@ func TestImportPaths(t *testing.T) {
 		fsys.On("ReadFile", "testdata/modules/imports.ts").Once()
 		fsys.On("ReadFile", "testdata/geometries/Geometries.js").Once()
 		// Run test
-		im := utils.ImportMap{}
+		im := ImportMap{}
 		err := im.WalkImportPaths("testdata/modules/imports.ts", fsys.ReadFile)
 		// Check error
 		assert.NoError(t, err)
@@ -68,12 +66,10 @@ func TestImportPaths(t *testing.T) {
 		fsys.On("ReadFile", "testdata/shared/mod.ts").Once()
 		fsys.On("ReadFile", "testdata/nested/index.ts").Once()
 		// Setup deno.json
-		im := utils.ImportMap{Imports: map[string]string{
+		im := ImportMap{Imports: map[string]string{
 			"module-name/": "../shared/",
 		}}
-		// Casting io.FS to afero.Fs
-		afsys := &afero.FromIOFS{FS: testImports}
-		assert.NoError(t, im.Resolve("testdata/modules/deno.json", afsys))
+		assert.NoError(t, im.Resolve("testdata/modules/deno.json", testImports))
 		// Run test
 		err := im.WalkImportPaths("testdata/modules/imports.ts", fsys.ReadFile)
 		// Check error
@@ -91,12 +87,10 @@ func TestImportPaths(t *testing.T) {
 		fsys.On("ReadFile", "testdata/shared/mod.ts").Once()
 		fsys.On("ReadFile", "testdata/nested/index.ts").Once()
 		// Setup legacy import map
-		im := utils.ImportMap{Imports: map[string]string{
+		im := ImportMap{Imports: map[string]string{
 			"module-name/": "./shared/",
 		}}
-		// Casting io.FS to afero.Fs
-		afsys := &afero.FromIOFS{FS: testImports}
-		assert.NoError(t, im.Resolve("testdata/import_map.json", afsys))
+		assert.NoError(t, im.Resolve("testdata/import_map.json", testImports))
 		// Run test
 		err := im.WalkImportPaths("testdata/modules/imports.ts", fsys.ReadFile)
 		// Check error
